Refactor dashboard
