+++ frontend/src/app/features/dashboard/components/awaiting-pickup/awaiting-pickup.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Dashboard } from '../dashboard';
+import { Dashboard } from '../../dashboard';
 import { AsyncPipe, NgIf } from '@angular/common';
-import { AppConsole } from '../../../utils/app-console';
+import { AppConsole } from '../../../../utils/app-console';
 
 @Component({
   selector: 'app-awaiting-pickup',
