+++ frontend/src/app/features/dashboard/components/recently-picked-up/recently-picked-up.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Dashboard } from '../dashboard';
+import { Dashboard } from '../../dashboard';
 import { AsyncPipe, NgIf } from '@angular/common';
 @Component({
   selector: 'app-recently-picked-up',
