+++ frontend/src/app/features/dashboard/dashboard-routing-module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DashboardParent } from './dashboard-parent/dashboard-parent';
+import { DashboardParent } from './pages/dashboard-admin/dashboard-parent';
 import { isLoggedInGuard, isAdminAndManagerAuthed } from '../../core/guards/auth-guard-guard';
 
 const routes: Routes = [
