+++ frontend/src/app/features/dashboard/pages/dashboard-admin/dashboard-parent.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { AwaitingPickup } from '../awaiting-pickup/awaiting-pickup';
-import { RecentlyPickedUp } from '../recently-picked-up/recently-picked-up';
+import { AwaitingPickup } from '../../components/awaiting-pickup/awaiting-pickup';
+import { RecentlyPickedUp } from '../../components/recently-picked-up/recently-picked-up';
 
 @Component({
   selector: 'app-dashboard-parent',
